Reference

IteratedIntegration.AbstractIteratedIntegrand — Type

AbstractIteratedIntegrand{d}

Supertype for integrands compatible with iterated integration of d variables.

(::AbstractIteratedIntegrand)(x)

Evaluate the integrand at the point x

IteratedIntegration.AbstractIteratedLimits — Type

AbstractIteratedLimits{d,T}

Supertype for limits that work with iterated integration

IteratedIntegration.CubicLimits — Type

CubicLimits(a, b)

Store integration limit information for a hypercube with vertices a and b. which can be can be real numbers, tuples, or AbstractVectors. The outermost variable of integration corresponds to the last entry.

IteratedIntegration.IteratedIntegrand — Type

IteratedIntegrand(fs...)

Represents a nested integral of the form ∫dxN fN( ... ∫dx1 f1([x1, ..., xN])) so the functions need to know the arguments and their layout, since the variables are passed as a vectors.

IteratedIntegration.PolyhedralLimits — Type

PolyhedralLimits(::Polyhedron)

Integration endpoints from a convex hull.

IteratedIntegration.ProductLimits — Type

ProductLimits(lims::AbstractIteratedLimits...)

Construct a collection of limits which yields the first limit followed by the second, and so on. The inner limits are not allowed to depend on the outer ones. The outermost variable of integration should be placed first, i.e. $\int_{\Omega} \int_{\Gamma}$ should be ProductLimits(Ω, Γ).

IteratedIntegration.TetrahedralLimits — Type

TetrahedralLimits(a::NTuple{d}) where d

A parametrization of the integration limits for a tetrahedron whose vertices are

( 0.0,  0.0, ...,  0.0)
( 0.0,  0.0, ..., a[d])
…
( 0.0, a[2], ..., a[d])
(a[1], a[2], ..., a[d])

IteratedIntegration.ThunkIntegrand — Type

ThunkIntegrand{d}(f)

Store f to evaluate f(x) at a later time. Employed by iterated_integration for generic integrands that haven't been specialized to use iterated_pre_eval. Note that x isa SVector, so the function arguments needs to expect the behavior. d is a parameter specifying the number of variables in the vector input of f. This is good for integrands like ∫∫∫f.

IteratedIntegration.TranslatedLimits — Type

TranslatedLimits(lims::AbstractIteratedLimits{d}, t::NTuple{d}) where d

Returns the limits of lims translated by offsets in t.

IntervalSets.endpoints — Function

endpoints(::AbstractLimits)

Return a tuple with the extrema of the next variable of integration. This is equivalent to projecting the integration domain onto one dimension.

IteratedIntegration.alloc_segbufs — Method

alloc_segbufs(coefficient_type, range_type, norm_type, ndim)
alloc_segbufs(coefficient_type, typesof_fx, typesof_nfx, ndim)
alloc_segbufs(f, l::AbstractIteratedLimits)

This helper function will allocate enough segment buffers as are needed for an nested_quadgk call of integrand f and integration limits l. coefficient_type should be eltype(l), typesof_fx should be the return type of the integrand f for each iteration of integration, typesof_nfx should be the types of the norms of a value of f for each iteration of integration, and ndim should be ndims(l).

IteratedIntegration.iai — Method

iai(f, a, b; kwargs...)
iai(f::AbstractIteratedIntegrand{d}, l::AbstractLimits{d}; order=7, atol=0, rtol=sqrt(eps()), norm=norm, maxevals=typemax(Int), segbuf=nothing) where d

Multi-dimensional globally-adaptive quadrature via iterated integration using Gauss-Kronrod rules. Interface is similar to nested_quadgk.

IteratedIntegration.iai_buffer — Method

iai_buffer(ndim, DT, RT, NT)

Allocate a buffer for use by iai for an integrand with integration over ndim dimensions with domain type DT, range type RT and norm type NT

IteratedIntegration.iterated_inference — Method

iterated_inference(F, T, ::Val{d}) where d
iterated_inference(f, l::AbstractIteratedLimits)

Returns a tuple of the return types of f after each variable of integration

IteratedIntegration.iterated_integral_type — Method

iterated_integral_type(f, l::AbstractIteratedLimits)

Returns the output type of nested_quadgk(f, l)

IteratedIntegration.iterated_integrand — Function

iterated_integrand(f::AbstractIteratedIntegrand{d}, x, ::Val{d}) where d

Evaluate a function on the inner integral x. When d==1, x is the argument of the innermost integral that is passed by the integration routine via f(x).

IteratedIntegration.iterated_pre_eval — Function

iterated_pre_eval(f::AbstractIteratedIntegrand{d}, x, ::Val{d}) where d

Perform a precomputation on f using the value of a variable of integration, x. Certain types of functions, such as Fourier series, take can use x to precompute a new integrand for the remaining variables of integration that is more computationally efficient. Otherwise, the type can store x and delay the evaluation to the inner integral. This function must return the integrand for the subsequent integral, which should be an AbstractIteratedIntegrand{d-1}

IteratedIntegration.iterated_segs — Method

iterated_segs(f, l::AbstractIteratedLimits, ::Val{initdivs}) where initdivs

Returns a Tuple of integration nodes that are passed to QuadGK to initialize the segments for adaptive integration. By default, returns initdivs equally spaced panels on endpoints(l). If f is localized, specializing this function can also help avoid errors when QuadGK fails to adapt.

IteratedIntegration.iterated_tol_update — Method

iterated_tol_update(f, l, atol, rtol)

Choose a new set of error tolerances for the next inner integral. By default returns (atol, rtol) unchanged.

IteratedIntegration.nested_quadgk — Method

nested_quadgk(f, a, b; kwargs...)
nested_quadgk(f::AbstractIteratedIntegrand{d}, ::AbstractIteratedLimits{d}; order=7, atol=nothing, rtol=nothing, norm=norm, maxevals=typemax(Int), initdivs=ntuple(i -> Val(1), Val{d}()), segbufs=nothing) where d

Calls QuadGK to perform iterated 1D integration of f over a compact domain parametrized by AbstractIteratedLimits. In the case two points a and b are passed, the integration region becomes the hypercube with those extremal vertices (which mimics hcubature). f is assumed to be type-stable.

Returns a tuple (I, E) of the estimated integral and estimated error.

Keyword options include a relative error tolerance rtol (if atol==0, defaults to sqrt(eps) in the precision of the norm of the return type), an absolute error tolerance atol (defaults to 0), a maximum number of function evaluations maxevals for each nested integral (defaults to 10^7), and the order of the integration rule (defaults to 7).

The algorithm is an adaptive Gauss-Kronrod integration technique: the integral in each interval is estimated using a Kronrod rule (2*order+1 points) and the error is estimated using an embedded Gauss rule (order points). The interval with the largest error is then subdivided into two intervals and the process is repeated until the desired error tolerance is achieved. This 1D procedure is applied recursively to each variable of integration in an order determined by l to obtain the multi-dimensional integral.

Unlike quadgk, this routine does not allow infinite limits of integration nor unions of intervals to avoid singular points of the integrand. However, the initdivs keyword allows passing a tuple of integers which specifies the initial number of panels in each quadgk call at each level of integration.

In normal usage, nested_quadgk will allocate segment buffers. You can instead pass a preallocated buffer allocated using alloc_segbufs as the segbuf argument. This buffer can be used across multiple calls to avoid repeated allocation.

Polyhedra.fixandeliminate — Function

fixandeliminate(l::AbstractIteratedLimits, x)

Fix the outermost variable of integration and return the inner limits.

For developers

Realizations of type T<:AbstractIteratedLimits only have to implement a method with signature fixandeliminate(::T, ::Number). The result must also have dimension one less than the input, and this should only be called when ndims

= 1