Reference

IteratedIntegration.AbstractIntegrator — Type

AbstractIntegrator{F} <: Function

Supertype of integration routines of a (collection of) function F.

IteratedIntegration.AbstractIteratedIntegrand — Type

AbstractIteratedIntegrand{d}

Supertype for integrands compatible with iterated integration of d variables.

IteratedIntegration.AbstractLimits — Type

AbstractLimits{d,T<:AbstractFloat}

Represents a set of integration limits over d variables of type T. Realizations of this type should implement lower and upper, which return the lower and upper limits of integration along some dimension, rescale which represents the number of symmetries of the BZ which are used by the realization to reduce the BZ (the integrand over the limits gets multiplied by this factor), and a functor that accepts a single numeric argument and returns another realization of that type (in order to do nested integration). Thus the realization is also in control of the order of variables of integration and must coordinate this behavior with their integrand. Instances should also be static structs.

IteratedIntegration.AssociativeOpIntegrand — Type

AssociativeOpIntegrand(op, I::AbstractIteratedIntegrand...)

Constructor for a collection of integrands reduced by an associative op.

IteratedIntegration.CubicLimits — Type

CubicLimits(a, b)

Store integration limit information for a hypercube with vertices a and b. which can be can be real numbers, tuples, or AbstractVectors. The outermost variable of integration corresponds to the last entry.

IteratedIntegration.IteratedIntegrand — Type

IteratedIntegrand(fs...; f0=identity)

Experimental

This may not work or may change

Represents a nested integral of the form f0(∫dxN fN(xN, ... ∫dx2 f2(x2, ..., xN, ∫dx1 f1(x1, ..., xN)) ... ))) so the functions need to know the arguments and their layout, since the variables and inner integrals are passed as vectors.

IteratedIntegration.IteratedIntegrator — Type

IteratedIntegrator(f, l, p...; ps=0.0, routine=iterated_integration, args=(), kwargs=(;))

Experimental

Intended to integrate all kinds of AbstractIteratedIntegrand

IteratedIntegration.PolyhedralLimits — Type

PolyhedralLimits(::Polyhedron)

Integration endpoints from a convex hull.

IteratedIntegration.ProductLimits — Type

ProductLimits(lims::AbstractLimits...)

Construct a collection of limits which yields the first limit followed by the second, and so on. The inner limits are not allowed to depend on the outer ones. The outermost variable of integration should be placed first, i.e. $\int_{\Omega} \int_{\Gamma}$ should be ProductLimits(Ω, Γ).

IteratedIntegration.TetrahedralLimits — Type

TetrahedralLimits(a::NTuple{d}) where d

A parametrization of the integration limits for a tetrahedron whose vertices are

( 0.0,  0.0, ...,  0.0)
( 0.0,  0.0, ..., a[d])
…
( 0.0, a[2], ..., a[d])
(a[1], a[2], ..., a[d])

IteratedIntegration.ThunkIntegrand — Type

ThunkIntegrand{d}(f)

Store f to evaluate f(x) at a later time. Employed by iterated_integration for generic integrands that haven't been specialized to use iterated_pre_eval. Note that x isa SVector, so the function arguments needs to expect the behavior. d is a parameter specifying the number of variables in the vector input of f. This is good for integrands like ∫∫∫f.

IteratedIntegration.TranslatedLimits — Type

TranslatedLimits(lims::AbstractLimits{d}, t::NTuple{d}) where d

Returns the limits of lims translated by offsets in t.

IntervalSets.endpoints — Method

endpoints(::AbstractLimits{d}, [dim=d]) where d

Return a tuple with the lower and upper limit of the dimth variable of integration, which is the outermost by default. This is equivalent to projecting the integration domain onto one dimension.

IteratedIntegration.alloc_segbufs — Method

alloc_segbufs(coefficient_type, typesof_fx, typesof_nfx, ndim)
alloc_segbufs(f, l::AbstractLimits)

This helper function will allocate enough segment buffers as are needed for an iterated_integration call of integrand f and integration limits l. coefficient_type should be coefficient_type(l), typesof_fx should be the return type of the integrand f for each iteration of integration, typesof_nfx should be the types of the norms of a value of f for each iteration of integration, and ndim should be ndims(l).

IteratedIntegration.iterated_inference — Method

iterated_inference(f, l::AbstractLimits{d})

Returns a tuple of the return types of f after each variable of integration

IteratedIntegration.iterated_integral_type — Method

iterated_integral_type(f, l)

Returns the output type of an iterated integral of f over domain l

IteratedIntegration.iterated_integrand — Method

iterated_integrand(f::AbstractIteratedIntegrand, y, ::Type{Val{d}}) where d
iterated_integrand(f::AbstractIteratedIntegrand, y, dim::Int) # fallback

By default, returns y which is the result of an interior integral. Can use dispatch on f to implement more kinds of iterated integrals, in which case the developer should know d takes values of 0, 1, ..., ndims(lims). d=1 evaluates the innermost integral, d=0 evaluates a function outside the last integral. Subtypes of AbstractFourierIntegrand

IteratedIntegration.iterated_integration — Method

iterated_integration(f, ::AbstractLimits; order=7, atol=nothing, rtol=nothing, norm=norm, maxevals=typemax(Int), initdivs=ntuple(i -> Val(1), Val{d}()), segbufs=nothing)
iterated_integration(f, a, b; kwargs...)

Calls QuadGK to perform iterated 1D integration of f over a compact domain parametrized by AbstractLimits. In the case two points a and b are passed, the integration region becomes the hypercube with those extremal vertices. f is assumed to be type-stable.

Returns a tuple (I, E) of the estimated integral and estimated error.

Keyword options include a relative error tolerance rtol (if atol==0, defaults to sqrt(eps) in the precision of the norm of the return type), an absolute error tolerance atol (defaults to 0), a maximum number of function evaluations maxevals for each nested integral (defaults to 10^7), and the order of the integration rule (defaults to 7).

The algorithm is an adaptive Gauss-Kronrod integration technique: the integral in each interval is estimated using a Kronrod rule (2*order+1 points) and the error is estimated using an embedded Gauss rule (order points). The interval with the largest error is then subdivided into two intervals and the process is repeated until the desired error tolerance is achieved. This 1D procedure is applied recursively to each variable of integration in an order determined by l to obtain the multi-dimensional integral.

Unlike quadgk, this routine does not allow infinite limits of integration nor unions of intervals to avoid singular points of the integrand. However, the initdivs keyword allows passing a tuple of integers which specifies the initial number of panels in each quadgk call at each level of integration.

In normal usage, iterated_integration will allocate segment buffers. You can instead pass a preallocated buffer allocated using alloc_segbufs as the segbuf argument. This buffer can be used across multiple calls to avoid repeated allocation.

IteratedIntegration.iterated_pre_eval — Method

iterated_pre_eval(f, x, dim)
iterated_pre_eval(f, x) # fallback

Perform a precomputation on f using the value of a variable of integration, x. The default is to store x and delay the computation of f(x) until all of the values of the variables of integration are determined at a integration point. Certain types of functions, such as Fourier series, take can use x to precompute a new integrand for the remaining variables of integration that is more computationally efficient. This function must return the integrand for the subsequent integral.

IteratedIntegration.iterated_segs — Method

iterated_segs(f, l::AbstractLimits, ::Val{initdivs}) where initdivs

Returns a Tuple of integration nodes that are passed to QuadGK to initialize the segments for adaptive integration. By default, returns initdivs equally spaced panels on endpoints(l). If f is localized, specializing this function can also help avoid errors when QuadGK fails to adapt.

IteratedIntegration.iterated_tol_update — Method

iterated_tol_update(f, l, atol, rtol)

Choose a new set of error tolerances for the next inner integral. By default returns (atol, rtol) unchanged.

IteratedIntegration.iterated_value — Method

iterated_value(f, Val{d}) where d

When d=1, calls iterated_value(f) to yield a pre-evaluated item from f to pass to an inner integral. This should be specialized for f

IteratedIntegration.quad_args — Method

quad_args(routine, f, lims, [...])

Return the tuple of arguments needed by the quadrature routine other than the test integrand f, depending on the test integrand f and lims.

IteratedIntegration.quad_kwargs — Method

quad_kwargs(routine, f, lims, kwargs::NamedTuple)

Supplies the default keyword arguments to the given integration routine without over-writing those already provided in kwargs

Polyhedra.coefficient_type — Method

coefficient_type(x::AbstractLimits)

Returns the type of the coordinate coefficients for the domain

Polyhedra.fixandeliminate — Function

fixandeliminate(l::AbstractLimits, x)

Fix the outermost variable of integration and return the inner limits.

For developers

Realizations of type T<:AbstractLimits only have to implement a method with signature fixandeliminate(::T, ::Number). The result must also have dimension one less than the input, and this should only be called when ndims

= 1